fix #27
for some reason, the build process wasn't copying the client files over
to the release directory like it should have been. this meant that none
of the client js was ending up in the built projects, which meant no
custom stuff was working. this fixes it, but not in a way that i'm
thrilled about.