check not only whether asset is streamed, but whether a remote URL is defined
[fix #20317673]

If no remote URL is defined, try to use the local URL even if the asset isn't
in the manifest; it may 404, but at least the 404 will be accurate.

This should have no effect on MAP-generated apps, as they should always have a url
defined for assets that don't appear in the manifest. This just handles the new
case presented by Mulberry, where it's possible for an asset to not have a rmote URL.