make build helper responsible for providing full template object
previously, the build helper only provided custom templates in the object
it returned; the generator was responsible for loading the base templates.
with this change, the build helper becomes responsible for providing all
templates for the app, and the base build helper is given a templates method
so that it will work "out of the box" with the base template set.

at the same time, the page templates generator is significantly dumbed down;
it now just returns a toura.templates string using whatever object is provided.
this is in keeping with the philosophy of app.rb in general -- it should just
format provided data as the app code requires.

this is all in support of #192, but also [re #20801221].