fix feed item list failure
it seems that since all of our i18n stuff is happening in the mulberry
namespace, it's important that this require happen inside a file that's
in them mulberry namespace in order for the required NLS files to end up
in the build. this is a non-optimal solution, but a better one will require
dealing with separating the Toura and Mulberry NLS stuff.